requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
dash>=2.9.3
pandas>=1.5.3
plotly>=5.14.1
